# expence-tracker/my-app/specs/1-expense-tracker-mvp/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Expense Tracker API
  description: RESTful API for the Expense Tracker MVP
  version: 1.0.0
  contact:
    name: API Support
    email: support@expensetracker.example.com
servers:
  - url: https://api.expensetracker.example.com/api
    description: Production server
  - url: http://localhost:3000/api
    description: Development server
...
paths:
...
  /auth/register:
    post:
      tags:
        - Authentication
      summary: Register new user
      description: Create a new user account with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                  maxLength: 255
                  example: user@example.com
                password:
                  type: string
                  format: password
                  minLength: 8
                  example: SecurePass123
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Verification email sent
                  userId:
                    type: string
                    format: uuid
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          description: Email already registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /auth/reset-password:
    post:
      tags:
        - Authentication
      summary: Request password reset
      description: Send password reset email to user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
              properties:
                email:
                  type: string
                  format: email
      responses:
        '200':
          description: Reset email sent (always returns 200 to prevent email enumeration)
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: If the email exists, a reset link has been sent
        '429':
          $ref: '#/components/responses/RateLimit'

  # Transactions
...
  /budgets:
    get:
      tags:
        - Budgets
      summary: List budgets
      description: Get budgets with spending progress
      security:
        - bearerAuth: []
      parameters:
        - name: month
          in: query
          description: Filter by month (YYYY-MM-01)
          schema:
            type: string
            format: date
      responses:
        '200':
          description: List of budgets with progress
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BudgetWithProgress'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - Budgets
      summary: Create budget
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BudgetInput'
      responses:
        '201':
          description: Budget created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Budget'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          description: Budget already exists for this category and month
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /budgets/{id}:
    put:
      tags:
        - Budgets
      summary: Update budget
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/BudgetId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                amount:
                  type: number
                  format: decimal
                  minimum: 0.01
                  maximum: 1000000
      responses:
        '200':
          description: Budget updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Budget'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags:
        - Budgets
      summary: Delete budget
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/BudgetId'
      responses:
        '204':
          description: Budget deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  # Reports
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    BudgetId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
  schemas:
...
    Category:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        type:
          type: string
          enum: [EXPENSE, INCOME]
        isPredefined:
          type: boolean
        createdAt:
          type: string
          format: date-time
...
    Budget:
      type: object
      properties:
        id:
          type: string
          format: uuid
        categoryId:
          type: string
          format: uuid
        category:
          $ref: '#/components/schemas/Category'
        amount:
          type: number
          format: decimal
        month:
          type: string
          format: date
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    BudgetInput:
      type: object
      required:
        - categoryId
        - amount
        - month
      properties:
        categoryId:
          type: string
          format: uuid
        amount:
          type: number
          format: decimal
          minimum: 0.01
          maximum: 1000000
        month:
          type: string
          format: date
          description: First day of the month (YYYY-MM-01)
    BudgetWithProgress:
      allOf:
        - $ref: '#/components/schemas/Budget'
        - type: object
          properties:
            spent:
              type: number
              format: decimal
            remaining:
              type: number
              format: decimal
            percentage:
              type: number
              format: float
              description: Spending as percentage of budget (0-100+)
            status:
              type: string
              enum: [SAFE, WARNING, EXCEEDED]
              description: SAFE (<70%), WARNING (70-100%), EXCEEDED (>100%)
...
    Error:
      type: object
      properties:
        error:
          type: string
        code:
          type: string
        details:
          type: object
          nullable: true
  responses:
    BadRequest:
      description: Invalid input
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Not authenticated
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    RateLimit:
      description: Rate limit exceeded
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
      headers:
        X-RateLimit-Limit:
          schema:
            type: integer
        X-RateLimit-Remaining:
          schema:
            type: integer
        X-RateLimit-Reset:
          schema:
            type: integer
            format: unix-timestamp